Update maker tests

+++ tests/conftest.py
@@ -1,7 +1,7 @@
 import pytest
 
 
-@pytest.fixture
+@pytest.fixture(scope="session")
 def test_dir():
     from pathlib import Path
 
@@ -10,12 +10,7 @@ def test_dir():
     return test_dir.resolve()
 
 
-@pytest.fixture
-def vasp_test_dir(test_dir):
-    return test_dir / "vasp"
-
-
-@pytest.fixture
+@pytest.fixture(scope="session")
 def log_to_stdout():
     import logging
     import sys
@@ -31,8 +26,8 @@ def log_to_stdout():
     root.addHandler(ch)
 
 
-@pytest.fixture
-def clean_dir():
+@pytest.fixture(scope="session")
+def clean_dir(debug_mode):
     import os
     import shutil
     import tempfile
@@ -41,5 +36,45 @@ def clean_dir():
     newpath = tempfile.mkdtemp()
     os.chdir(newpath)
     yield
-    os.chdir(old_cwd)
-    shutil.rmtree(newpath)
+    if debug_mode:
+        print(f"Tests ran in {newpath}")
+    else:
+        os.chdir(old_cwd)
+        shutil.rmtree(newpath)
+
+
+@pytest.fixture(scope="session")
+def debug_mode():
+    return False
+
+
+@pytest.fixture(scope="session")
+def lpad(database, debug_mode):
+    from fireworks import LaunchPad
+
+    lpad = LaunchPad(name=database)
+    lpad.reset("", require_password=False)
+    yield lpad
+
+    if not debug_mode:
+        lpad.reset("", require_password=False)
+        for coll in lpad.db.list_collection_names():
+            lpad.db[coll].drop()
+
+
+@pytest.fixture(scope="function")
+def memory_jobstore():
+    from jobflow import JobStore
+    from maggma.stores import MemoryStore
+
+    store = JobStore(MemoryStore())
+    store.connect()
+
+    return store
+
+
+@pytest.fixture(scope="session", autouse=True)
+def log_to_stdout():
+    from atomate2.utils.log import initialize_logger
+
+    initialize_logger()
